fix: remove 'package-json' definition
Observed that it only behaves as intended when there is no definition
for it in utils/config/definitions.js. It seems that @npmcli/config (or
nopt) gets confused when using the following configuration:

  default: null,
  type: [null, true, String],

In particular, if no path argument is given with --package-json, but it
is followed by another --option, the '--option' gets treated as the
path argument to --package-json. This is an aberration beyond the
control of this project.