add handling to skip alias spec in download stage
Given the ability to use a package.json to supply download specs, and
that the user may have an alias spec in the package.json and neglect to
remove it for download purposes, it would be unreasonable for the
download of any other listed packages to fail when we can just ignore
alias specs. Additionally, a warning is logged.