refactored package-json option
Restored the definition, but simplified to type String. Not using type
path, because @npmcli/config resolves such args to absolute paths, which
is not needed here. @npmcli/config has other aberrant behaviors, but
those are now gracefully handled for this option in download.js.
Refactored tests accordingly, and added a feature in mock item-agents to
allow checking the options that were passed.