added read-from-tarball, test and test assets
First step towards refactoring download behavior: re-purpose the
untar-to-memory idea. We can't get lock file information from the
manifest of a package; we need to fetch the tarball, look inside
for a lock file, and use that to determine dependencies; in absence
of a lock file, fall back to the manifest.