add alt-yarn-lock and corresponding test files
arborist yarn-lock has a flawed approach to 'resolved' values that have
hash suffixes: it removes them, regardless of whether they are the
"legacy shasum hashes" that yarn apparently added to registry URLs in
times gone by, or commit-ish values that are necessary to identify
specific commits for git repo specs.

alt-yarn-lock.js has a correction for this, and only contains the
features of arborist yarn-lock.js that are necessary for parsing the
content of a yarn.lock file.