added function readFromDir to lock-deps.js
Having developed all that's needed to get dependencies from both kinds
of lockfiles, then going on to refactoring the download stage to handle
these from tarballs, it seemed a reasonable idea to add a command line
feature similar to the one for a package.json file. readFromDir is the
basis for this. Had to factor it out of download.js and into
lock-deps so that it can be mocked in the download_test, otherwise we
are faced with trying to mock fs.readFile in a test suite where we are
already relying on actual fs.readFile.