Full rewrite of fromYarnLock()
If we must use yarn.lock files sometimes, then we must do it right:
the output of fromYarnLock must correspond as closely as possible to
that of fromPackageLock for the same package installation. That means
reconstructing the heirarchy starting from the package.json, following
the dependency relationships down to the leaf packages, categorizing
each item as we go, and then adding 'peer', 'optional', 'dev', and
even 'devOptional' flags where applicable.
The cost of not doing the above, in particular, is that we end up
fetching loads of devDependencies when they're not requested, simply
because a yarn.lock (v1) cannot tell the difference.

Also expanded extract() and readFromDir() to handle the need to fetch
a package.json in the yarn.lock case.