added 2 pkgs to registry-mocks content
REALLY wanted to avoid changes to the registry-mocks content, because it
is borrowed from arborist test fixtures, so it would be easiest to
manage future versions of npm-two-stage if that remains unchanged.
However, more fixtures were needed for a more nuanced test of a
dependency package with a yarn.lock in it.