fix: (dev) added retry to audit in add()
In stress test on Windows, `npm download` a package with many many dependencies,
frequently saw EZEROLEN error. When command is repeated with same target, error
does not reoccur.
Hypothesis: filesystem is sluggish, especially when malware protection is running.

Also changed 'Could not find a map file' log message to 'info' level, so
that it does not show at the default log level.