fix: added retry to ops on files post-download
Occasional failures on trying to read tarballs during npm download, on
Windows, when the dependency tree is very large (e.g., tap@16) led to
the discovery that the promise returned by pacote.tarball.file()
sometimes resolves *before the file has been written to the disk.*
Treated in 2 places: on attempt to extract a lockfile; and on adding a
package record to the dl tracker (because add() calls auditOne(), which
calls fs.lstat and checks the file size).
This fix qualifies as a Hail Mary.