fix: handle fs error on rm of temp directory
Windows sometimes gives an EPERM, depending on what's in the temp
directory. In the one case seen, it was a git repo clone.
Expect that EBUSY may also happen occasionally. In fact, handle any
error from rm in that final clause, because at that point we have all
the tarballs. Just make sure that dltracker.serialize is called first.
There's no good reason to fail the whole operation after that.
Give the user a helpful message if rm still fails after repeated
attempts with a backoff.