test: don't touch outside of test directory
Create a etc/npmrc in the staging location;
use 'logs-dir' to direct logging to a temp directory inside staging;
use 'update-notifier=false' to suppress the update notification.
Pass the option to set the location of the cache inside a tap.testdir
for 'npm install' calls.
Without these measures, the staged npm was writing stuff to the user's
production cache directory.