fix: ignore lockfile for offline install
Temporary solution to the problem: on determining offline state, set
'package-lock' option false before passing options to AltArborist. This
ensures that offline installation is driven only by package.json. This
will be consistent with lockfile-driven behavior as long as there are
no updates to the tarball directory used as a source for successive
installations.

A long-term solution would be in build-ideal-tree.js, would derive
key data from the 'resolved' value (which comes from a lockfile), and
would get the tarball filename from the dltracker and put it in the map
so that it is available to reify.js when needed. Then lockfiles would
not have to be disabled.