fix: get normalized JSON errors
JSON built-in gives different error messages depending on node.js
version. This causes the lock-deps test to fail for a trivial reason
when run on a differing version of node.js.
Use json-parse-even-better-errors, which is bundled with npm; then we
can rely on the error message content we know we get with the version
that is bundled with the targeted version of npm.