test: follow-up: get normalized JSON errors
Can't even trust json-parse-even-better-errors.
Given an 'Unexpected token' case, there is a difference in its output
depending on whether it's running on node.js less-than-or-eq-to 18.16.0
or greater-than-or-eq-to 19.9.0: the first unexpected token is wrapped
in double quotes on the former, but single quotes on the latter.
So we generalize our expectations in the test.
No change to the production code this time.