fix: get normalized JSON errors
JSON built-in gives different error messages depending on node.js
version. This causes the lock-deps test to fail for a trivial reason
when run on a differing version of node.js.
Use json-parse-even-better-errors, which is bundled with npm; then we
should be able to rely on the error message content we get with the
version that is bundled with the targeted version of npm.

Addendum: Can't even trust json-parse-even-better-errors.
Given an 'Unexpected token' case, there is a difference in its output
depending on whether it's running on node.js less-than-or-eq-to 18.16.0
or greater-than-or-eq-to 19.9.0: the first unexpected token is wrapped
in double quotes on the former, but single quotes on the latter.
So we generalize our expectations in the test.