fix: get normalized JSON errors
JSON built-in gives different error messages depending on node.js
version. This causes the lock-deps test to fail for a trivial reason
when run on a differing version of node.js.
Use json-parse-even-better-errors, which is bundled with npm; then we
should be able to rely on the error message content we get with the
version that is bundled with the targeted version of npm.

The above explanation is moot while this version is still using
tap 15, because that does not run on node.js > 16. Propagated the
changes anyway just to be consistent with higher-numbered versions that
have already received this change.